move to chain.rs
